逻辑大体实现

+++ src/main/java/com/donler/gym/controller/BusinessController.java
@@ -1,7 +1,12 @@
 package com.donler.gym.controller;
 
+import com.donler.gym.expection.AttrValidateException;
 import com.donler.gym.model.Business;
+import com.donler.gym.model.dto.DeleteStatusModel;
 import com.donler.gym.repo.BusinessRepo;
+import com.donler.gym.util.NullCheckUtils;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,32 +19,54 @@ import java.util.List;
  */
 @RestController
 @RequestMapping(value = "/business")
+@Api(value = "business", tags = {"项目"})
 public class BusinessController {
 
 
   @Autowired
   private BusinessRepo businessRepo;
 
+  @ApiOperation(value = "添加一个项目选项", notes = "通过传入的选项信息生成一个项目选项")
   @RequestMapping(value = "", method = RequestMethod.POST, produces = {"application/json"}, consumes = {"application/json"})
-  public ResponseEntity<Business> addBusiness(@RequestHeader(value = "x-token",required = true) String token, @Valid @RequestBody Business input) {
+  public ResponseEntity<Business> addBusiness(@Valid @RequestBody Business input) {
     Business business = businessRepo.save(input);
     return ResponseEntity.ok(business);
   }
 
-  @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = {"application/json"}, consumes = {"application/json"})
-  public String deleteBusiness(@PathVariable("id") Long id) {
+  @ApiOperation(value = "删除指定的项目选项", notes = "删除指定的项目选项,并给出反馈")
+  @RequestMapping(value = "/{id}", method = RequestMethod.DELETE, produces = {"application/json"})
+  public ResponseEntity<DeleteStatusModel> deleteBusiness(@PathVariable("id") Long id) {
+    Business business = businessRepo.findOne(id);
+    if (NullCheckUtils.isNullOrEmpty(business)) {
+      throw new AttrValidateException("请传入正确的id");
+    }
     businessRepo.delete(id);
-    return "delete success";
+
+    return ResponseEntity.ok(new DeleteStatusModel());
   }
 
-  @RequestMapping(value = "", method = RequestMethod.GET, produces = {"application/json"}, consumes = {"application/json"})
+  @ApiOperation(value = "获取所有项目选项", notes = "获得所有的项目选项列表")
+  @RequestMapping(value = "", method = RequestMethod.GET, produces = {"application/json"})
   public ResponseEntity<List<Business>> findAllBusiness() {
     return ResponseEntity.ok(businessRepo.findAll());
   }
 
+
+  @ApiOperation(value = "更新指定项目选项", notes = "传入项目选项的信息更新指定的项目选项(ps.传入的实体中的id忽略,后台不予以识别)")
   @RequestMapping(value = "/{id}", method = RequestMethod.PUT, produces = {"application/json"}, consumes = {"application/json"})
-  public ResponseEntity<Business> updateBusiness(@PathVariable("id") Long id,@Valid @RequestBody Business input) {
-    return ResponseEntity.ok(businessRepo.save(input));
+  public ResponseEntity<Business> updateBusiness(@PathVariable("id") Long id, @Valid @RequestBody Business input) {
+    Business business = businessRepo.findOne(id);
+    if (NullCheckUtils.isNullOrEmpty(business)) {
+      throw new AttrValidateException("请检查传入的id");
+    }
+
+    /**
+     * 更新操作
+     */
+    business.setPrice(input.getPrice());
+    business.setMonthCount(input.getMonthCount());
+
+    return ResponseEntity.ok(businessRepo.save(business));
   }
 
 }
